feat: Add external product link support
Introduces the `externalUrl` field to the `Product` type and integrates it into the `ProductCard` component.
When an `externalUrl` is present, clicking the action button will now open this URL in a new tab, providing an alternative purchase option. Otherwise, the existing "add to cart" functionality remains.
Also adds a sample `externalUrl` to a product in `constants.ts`.